Include `user_id` reference column in all tables
When querying for any data in our application, only the data that belongs
to the logged-in user should be returned to them. This is most important
in GET requests for our various resources, notable families, children,
and chores. We wouldn't want to return chores that belong to a different
family to the user.

With the original database design, we were going to have to use subque-
ries in order to find the resources that belong to a specific user. To
make things more simple, this change adds a `user_id` reference column
to all tables, eliminating the need for complex subqueries.